+++ packages/client/src/__tests__/server-side/call.test.ts
@@ -120,6 +120,7 @@ describe('call API', () => {
       response = await call.update({
         settings_override: {
           recording: {
+            mode: RecordSettingsRequestModeEnum.AVAILABLE,
             audio_only: false,
             quality: RecordSettingsRequestQualityEnum._1080P,
             mode: RecordSettingsRequestModeEnum.AUTO_ON,