+++ sample-apps/react/egress-composite/src/components/layouts/DominantSpeaker/DominantSpeaker.tsx
@@ -7,8 +7,9 @@ import {
 } from '@stream-io/video-react-sdk';
 import { useSpotlightParticipant } from './useSpotlightParticipant';
 import { useEgressReadyWhenAnyParticipantMounts } from '../egressReady';
-import './DominantSpeaker.scss';
 import { AudioTracks } from './AudioTracks';
+
+import './DominantSpeaker.scss';
 
 export const DominantSpeaker = () => {
   const activeCall = useCall();
@@ -24,7 +25,10 @@ export const DominantSpeaker = () => {
   if (!activeCall) return <h2>No active call</h2>;
   return (
     <>
-      <div className="dominant-speaker__container">
+      <div
+        className="dominant-speaker__container"
+        data-testid="single_participant"
+      >
         {speakerInSpotlight && (
           <ParticipantView
             participant={speakerInSpotlight}