+++ sample-apps/react/egress-composite/src/components/layouts/Spotlight/Spotlight.tsx
@@ -12,7 +12,7 @@ export const Spotlight = () => {
   } = useConfigurationContext();
 
   return (
-    <div className="spotlight">
+    <div className="spotlight" data-testid="spotlight">
       <SpeakerLayout
         participantsBarPosition={barPosition}
         ParticipantViewUIBar={