+++ packages/react-native-sdk/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+### [0.0.18](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.17...@stream-io/video-react-native-sdk-0.0.18) (2023-09-19)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `0.1.0`
+* `@stream-io/video-react-bindings` updated to version `0.2.21`
 ### [0.0.17](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.16...@stream-io/video-react-native-sdk-0.0.17) (2023-09-18)
 
 ### Dependency Updates